Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.40.2
google-genai==0.8.0
